Move `debugbuild`
