+++ internal/util/debugbuild/debugbuild_on.go
@@ -1,0 +1,24 @@
+// Copyright 2021 FerretDB Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build ferretdb_debug || ferretdb_testcover || race
+
+package debugbuild
+
+// Enabled is true if that's a debug build.
+//
+// See build/version package documentation for more details.
+//
+// It is a constant to allow the compiler to optimize away the code.
+const Enabled = true
